data_services/requirements.txt
paho-mqtt
requests
AWSIoTPythonSDK
APScheduler

